+++ src/cyclebane/graph.py
@@ -2,16 +2,13 @@
 # Copyright (c) 2024 Scipp contributors (https://github.com/scipp)
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
-from collections import abc
 from dataclasses import dataclass
 from typing import Any, Generator, Hashable, Iterable, Mapping, Sequence
 from uuid import uuid4
 
 import networkx as nx
 
-IndexName = Hashable
-IndexValue = Hashable
+from .node_values import IndexName, IndexValue, NodeValues
 
 
 def _is_pandas_series_or_dataframe(obj: Any) -> bool:
@@ -257,241 +254,6 @@ class PositionalIndexer:
 
 
 MappingToArrayLike = Any  # dict[str, Numpy|DataArray], DataFrame, etc.
-
-
-class ValueArray(ABC):
-    """A series of values with an index that can be sliced."""
-
-    @staticmethod
-    def from_array_like(values: Any, *, axis_zero: int = 0) -> ValueArray:
-        if hasattr(values, 'dims'):
-            return DataArrayAdapter(values)
-        if values.__class__.__name__ == 'ndarray':
-            return NumpyArrayAdapter(values, axis_zero=axis_zero)
-        return IterableAdapter(values, index=range(len(values)), axis_zero=axis_zero)
-
-    @abstractmethod
-    def isel(self, key: list[tuple[IndexName, IndexValue]]) -> Any:
-        pass
-
-    @abstractmethod
-    def __getitem__(self, key: int | slice | tuple[int | slice, ...]) -> ValueArray:
-        pass
-
-    @property
-    @abstractmethod
-    def shape(self) -> tuple[int, ...]:
-        pass
-
-    @property
-    @abstractmethod
-    def index_names(self) -> tuple[IndexName, ...]:
-        pass
-
-    @property
-    @abstractmethod
-    def indices(self) -> dict[IndexName, Iterable[IndexValue]]:
-        pass
-
-
-class PandasSeriesAdapter(ValueArray):
-    def __init__(self, series, *, axis_zero: int = 0):
-        self._series = series
-        self._axis_zero = axis_zero
-
-    def isel(self, key: list[tuple[IndexName, IndexValue]]) -> Any:
-        if len(key) != 1:
-            raise ValueError('PandasSeriesAdapter only supports single index')
-        _, i = key[0]
-        return self._series.loc[i]
-
-    def __getitem__(
-        self, key: int | slice | tuple[int | slice, ...]
-    ) -> PandasSeriesAdapter:
-        return PandasSeriesAdapter(self._series[key])
-
-    @property
-    def shape(self) -> tuple[int, ...]:
-        return (len(self._series),)
-
-    @property
-    def index_names(self) -> tuple[IndexName, ...]:
-        return (self._series.index.name,)
-
-    @property
-    def indices(self) -> dict[IndexName, Iterable[IndexValue]]:
-        name = (
-            self._series.index.name
-            if self._series.index.name is not None
-            else f'dim_{self._axis_zero}'
-        )
-        return {name: self._series.index}
-
-
-class DataArrayAdapter(ValueArray):
-    def __init__(self, data_array):
-        self._data_array = data_array
-
-    def isel(self, key: list[tuple[IndexName, IndexValue]]) -> Any:
-        if hasattr(self._data_array, 'isel'):
-            return self._data_array.isel(dict(key))
-        values = self._data_array
-        for label, i in key:
-            # This is Scipp notation, Xarray uses the 'isel' method.
-            values = values[(label, i)]
-        return values
-
-    def __getitem__(
-        self, key: int | slice | tuple[int | slice, ...]
-    ) -> DataArrayAdapter:
-        return DataArrayAdapter(self._data_array[key])
-
-    @property
-    def shape(self) -> tuple[int, ...]:
-        return self._data_array.shape
-
-    @property
-    def index_names(self) -> tuple[IndexName, ...]:
-        return tuple(self._data_array.dims)
-
-    @property
-    def indices(self) -> dict[IndexName, Iterable[IndexValue]]:
-        # TODO Cannot be range after slicing!
-        return {name: range(size) for name, size in zip(self.index_names, self.shape)}
-
-
-class NumpyArrayAdapter(ValueArray):
-    def __init__(self, array, *, axis_zero: int = 0):
-        import numpy as np
-
-        self._array = np.asarray(array)
-        self._axis_zero = axis_zero
-
-    def isel(self, key: list[tuple[IndexName, IndexValue]]) -> Any:
-        return self._array[tuple(i for _, i in key)]
-
-    def __getitem__(
-        self, key: int | slice | tuple[int | slice, ...]
-    ) -> NumpyArrayAdapter:
-        return NumpyArrayAdapter(self._array[key], axis_zero=self._axis_zero)
-
-    @property
-    def shape(self) -> tuple[int, ...]:
-        return self._array.shape
-
-    @property
-    def index_names(self) -> tuple[IndexName, ...]:
-        return tuple(f'dim_{i+self._axis_zero}' for i in range(self._array.ndim))
-
-    @property
-    def indices(self) -> dict[IndexName, Iterable[IndexValue]]:
-        # TODO Cannot be range after slicing!
-        return {name: range(size) for name, size in zip(self.index_names, self.shape)}
-
-
-class IterableAdapter(ValueArray):
-    def __init__(
-        self, values: Sequence[Any], *, index: tuple[IndexValue], axis_zero: int = 0
-    ):
-        self._values = values
-        self._index = index
-        self._axis_zero = axis_zero
-
-    def isel(self, key: list[tuple[IndexName, IndexValue]]) -> Any:
-        if len(key) != 1:
-            raise ValueError('IterableAdapter only supports single index')
-        _, i = key[0]
-        return self._values[self._index.index(i)]
-
-    def __getitem__(
-        self, key: int | slice | tuple[int | slice, ...]
-    ) -> IterableAdapter:
-        return IterableAdapter(
-            self._values[key], index=self._index, axis_zero=self._axis_zero
-        )
-
-    @property
-    def shape(self) -> tuple[int, ...]:
-        return (len(self._values),)
-
-    @property
-    def index_names(self) -> tuple[IndexName, ...]:
-        return (f'dim_{self._axis_zero}',)
-
-    @property
-    def indices(self) -> dict[IndexName, Iterable[IndexValue]]:
-        return {f'dim_{self._axis_zero}': self._index}
-
-
-class NodeValues(abc.Mapping):
-    """A collection of pandas.DataFrame-like objects with distinct indices."""
-
-    def __init__(self, values: Mapping[Hashable, ValueArray]):
-        self._values = values
-
-    def __len__(self) -> int:
-        """Return the number of columns."""
-        return len(self._values)
-
-    def __iter__(self) -> Iterable[Hashable]:
-        """Iterate over the column names."""
-        return iter(self._values)
-
-    def __getitem__(self, key: Hashable) -> ValueArray:
-        """Return the column with the given name."""
-        return self._values[key]
-
-    def _to_value_arrays(
-        self, values: Mapping[Hashable, Sequence[Any]]
-    ) -> Mapping[Hashable, ValueArray]:
-        keys = tuple(values)
-        ndim = len(self.indices)
-        if (columns := getattr(values, 'columns', None)) is not None:
-            return {
-                key: PandasSeriesAdapter(values.iloc[:, i], axis_zero=ndim)
-                for key, i in zip(keys, range(len(columns)))
-            }
-        return {
-            key: ValueArray.from_array_like(values[key], axis_zero=ndim) for key in keys
-        }
-
-    def merge_from_mapping(
-        self, node_values: Mapping[Hashable, Sequence[Any]]
-    ) -> NodeValues:
-        """Append from a mapping of node names to value sequences."""
-        for node in node_values:
-            if any(node in mapping for mapping in self._values.keys()):
-                raise ValueError(f"Node '{node}' has already been mapped")
-        value_arrays = self._to_value_arrays(node_values)
-        shapes = [array.shape for array in value_arrays.values()]
-        if len(set(shapes)) != 1:
-            raise ValueError(
-                'All value sequences in a map operation must have the same shape. '
-                'Use multiple map operations if necessary.'
-            )
-        return self.merge(value_arrays)
-
-    def merge(self, value_arrays: Mapping[Hashable, ValueArray]) -> NodeValues:
-        if value_arrays:
-            named = next(iter(value_arrays.values())).index_names
-            if any([name in self.indices for name in named]):
-                raise ValueError(
-                    f'Conflicting new index names {named} with existing '
-                    f'{tuple(self.indices)}'
-                )
-        return NodeValues({**self._values, **value_arrays})
-
-    def get_columns(self, keys: list[Hashable]) -> NodeValues:
-        """Select a subset of columns."""
-        return NodeValues({key: self._values[key] for key in keys})
-
-    @property
-    def indices(self) -> dict[IndexName, Iterable[IndexValue]]:
-        """Return the indices of the NodeValues object."""
-        value_indices = [value.indices for value in self._values.values()]
-        return {
-            name: index for indices in value_indices for name, index in indices.items()
-        }
 
 
 class Graph:
